Move the config file scheme into a separate package so it can be imported by the operator.
Move the argument for the number of servers per pod into a command-line argument so we can use the same config file at different storage server densities.
