+++ fdbkubernetesmonitor/api/.testdata/default_config.json
@@ -1,6 +1,5 @@
 {
 	"version": "6.3.15",
-	"serverCount": 1,
 	"arguments": [
 		{"value": "--cluster_file"},
 		{"value": ".testdata/fdb.cluster"},
