+++ fdbkubernetesmonitor/api/config_test.go
@@ -17,7 +17,7 @@
 // limitations under the License.
 //
 
-package main
+package api
 
 import (
 	"encoding/json"
